chore(parser): move ingressRules to ingressrules.go
